rename abstractmodel
