+++ abstract_model.py
